Moved general stream processing functions out of RF module into parent directory.
