+++ seismic/units_utils.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+"""
+Constants and utility functions for unit conversion.
+"""
+
+KM_PER_DEG = 111.1949
